+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 
 [project]
-name = "abcd-microstructure-pipelines"
+name = "abcdmicro"
 authors = [
   { name = "Ebrahim Ebrahim", email = "ebrahim.ebrahim@kitware.com" },
   { name = "David Allemang", email = "david.allemang@kitware.com" },
@@ -61,10 +61,10 @@ Discussions = "https://github.com/brain-microstructure-exploration-tools/abcd-mi
 Changelog = "https://github.com/brain-microstructure-exploration-tools/abcd-microstructure-pipelines/releases"
 
 [project.scripts]
-gen_masks = "abcd_microstructure_pipelines.run:gen_masks"
+gen_masks = "abcdmicro.run:gen_masks"
 
 [tool.setuptools_scm]
-write_to = "src/abcd_microstructure_pipelines/_version.py"
+write_to = "src/abcdmicro/_version.py"
 
 
 [tool.pytest.ini_options]
@@ -81,7 +81,7 @@ testpaths = [
 
 
 [tool.coverage]
-run.source = ["abcd_microstructure_pipelines"]
+run.source = ["abcdmicro"]
 report.exclude_also = [
   '\.\.\.',
   'if typing.TYPE_CHECKING:',
@@ -98,7 +98,7 @@ disallow_untyped_defs = false
 disallow_incomplete_defs = false
 
 [[tool.mypy.overrides]]
-module = "abcd_microstructure_pipelines.*"
+module = "abcdmicro.*"
 disallow_untyped_defs = true
 disallow_incomplete_defs = true
 
@@ -137,7 +137,7 @@ ignore = [
 ]
 isort.required-imports = ["from __future__ import annotations"]
 # Uncomment if using a _compat.typing backport
-# typing-modules = ["abcd_microstructure_pipelines._compat.typing"]
+# typing-modules = ["abcdmicro._compat.typing"]
 
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = ["T20"]
