+++ tests/test_package.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import importlib.metadata
 
-import abcd_microstructure_pipelines as m
+import abcdmicro as m
 
 
 def test_version():
-    assert importlib.metadata.version("abcd_microstructure_pipelines") == m.__version__
+    assert importlib.metadata.version("abcdmicro") == m.__version__
